Refactor code into custom components and hooks
